refactor(hub): split mcp.rs into args/codecs/tools submodules
Before: one 2017-line `mcp.rs` mixed request/response shapes, the
#[tool_router] impl, codec helpers, state wiring, and a 1150-line
test module.

After:
- `mcp/args.rs` (263 lines): every `*Args` request struct and every
  response type (`MailStatus`, `EngineInfo`, `SpawnResult`, etc.),
  plus the tiny `one()` and `log_level_to_str()` helpers and the
  `CaptureFrameResultWire` decode mirror. Pure data, no logic.
- `mcp/codecs.rs` (143 lines): `deliver_one`, `resolve_payload`,
  `find_kind`, `encode_capture_bundle`, `decode_inbound` — the
  schema-driven encode/decode path between tool JSON and wire bytes.
- `mcp/tools.rs` (426 lines): the `#[tool_router] impl Hub` block
  with all eight tool methods, the `ServerHandler` impl, and
  `Hub::new` (moved here so it can call the macro-generated private
  `Self::tool_router()`). Capture-specific constants are scoped here
  too since nothing else references them.
- `mcp.rs` (root, ~130 non-test lines): `HubState`, `Hub` struct,
  `run_mcp_server`, and the untouched `#[cfg(test)] mod tests` block.

Test module is preserved verbatim. Test-only imports from protocol
and args are gated on `#[cfg(test)]` at the root so `use super::*;`
in the test module keeps working without touching individual tests.
Tool methods are `pub(super)` so the test module can still call them
directly.

Public API (`lib.rs` re-exports `DEFAULT_MCP_PORT`, `HubState`,
`run_mcp_server`) is unchanged.

No behavior change. All 111 hub unit tests pass.
